Changes for reading image from passed DB
Took the passed DB from findDataBox and wrote a method to sharpen it. fount the rough size of the edges and took that into account when iterating through the cells. I think part of the issue lies with the fact that we are taking a picture of a screen. The pixels are showing up in the camera which messes with the color recognition.